Allocate processor file
